# shared/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Stellive Notification Hub API
  version: 0.1.0
  description: Unofficial fan-project API. Official YouTube live events for the Stellive official channel are excluded.
paths:
...
  /v1/admin/hub-events:
    get:
      summary: List admin hub events
      responses:
        "200":
          description: Admin hub event list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminHubEventListResponse"
        "401":
          description: Admin authentication required.
    post:
      summary: Create a draft hub event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminHubEventCreateRequest"
      responses:
        "201":
          description: Draft hub event created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminHubEventResponse"
        "400":
          description: Validation failed.
        "401":
          description: Admin authentication required.
  /v1/admin/hub-events/{id}:
    get:
      summary: Get an admin hub event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Admin hub event detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminHubEventResponse"
        "404":
          description: Hub event not found.
    put:
      summary: Update an admin hub event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminHubEventUpdateRequest"
      responses:
        "200":
          description: Hub event updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminHubEventResponse"
        "400":
          description: Validation failed.
    delete:
      summary: Soft delete an admin hub event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hub event soft deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminHubEventResponse"
  /v1/admin/hub-events/{id}/publish:
    post:
      summary: Publish a hub event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hub event published.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminHubEventResponse"
  /v1/admin/hub-events/{id}/schedule-items:
    post:
      summary: Add a schedule item to an admin hub event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminHubEventScheduleItemCreateRequest"
      responses:
        "201":
          description: Schedule item created and parent revision incremented.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminHubEventResponse"
        "409":
          description: Parent revision conflict.
  /v1/admin/hub-events/{id}/schedule-items/{scheduleItemId}:
    patch:
      summary: Update a schedule item
      parameters:
        - $ref: "#/components/parameters/HubEventId"
        - $ref: "#/components/parameters/HubEventScheduleItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminHubEventScheduleItemMutationRequest"
      responses:
        "200":
          description: Schedule item updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminHubEventResponse"
        "409":
          description: Parent revision conflict.
    delete:
      summary: Delete a safe draft schedule item or cancel a reference-prone item
      parameters:
        - $ref: "#/components/parameters/HubEventId"
        - $ref: "#/components/parameters/HubEventScheduleItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminHubEventRevisionRequest"
      responses:
        "200":
          description: Schedule item deleted or cancelled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminHubEventScheduleDeleteResponse"
        "409":
          description: Parent revision conflict.
  /v1/admin/hub-events/{id}/schedule-items/{scheduleItemId}/restore:
    post:
      summary: Restore a cancelled schedule item
      parameters:
        - $ref: "#/components/parameters/HubEventId"
        - $ref: "#/components/parameters/HubEventScheduleItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminHubEventRevisionRequest"
      responses:
        "200":
          description: Schedule item restored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminHubEventResponse"
        "409":
          description: Parent revision conflict.
  /v1/admin/hub-events/{id}/schedule-items/order:
    put:
      summary: Reorder all schedule items
      parameters:
        - $ref: "#/components/parameters/HubEventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminHubEventScheduleOrderRequest"
      responses:
        "200":
          description: Schedule items reordered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminHubEventResponse"
        "409":
          description: Parent revision conflict.
  /v1/admin/hub-events/{id}/cancel:
    post:
      summary: Cancel a published hub event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hub event cancelled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminHubEventResponse"
  /v1/admin/hub-events/{id}/deactivate:
    post:
      summary: Deactivate a hub event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hub event deactivated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminHubEventResponse"
  /v1/admin/hub-events/{id}/audit-log:
    get:
      summary: List hub event audit log entries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hub event audit log.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/HubEventAuditLogEntry"
  /v1/admin/hub-events/validate:
    post:
      summary: Validate a hub event write payload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - input
              properties:
                mode:
                  type: string
                  enum: [draft, publish]
                  default: draft
                input:
                  $ref: "#/components/schemas/AdminHubEventCreateRequest"
      responses:
        "200":
          description: Validation result.
          content:
            application/json:
              schema:
                type: object
                required:
                  - valid
                  - errors
                properties:
                  valid:
                    type: boolean
                  errors:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminHubEventValidationError"
...
components:
  securitySchemes:
    internalBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Bearer token protection for `/v1/internal/*` operational routes.
  parameters:
    HubEventId:
      name: id
      in: path
      required: true
      schema:
        type: string
    HubEventScheduleItemId:
      name: scheduleItemId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    HubEventImagePolicyState:
      type: string
      enum:
        - none
        - official_runtime_url
        - third_party_allowed
        - verify_required
        - blocked
    HubEventImage:
      type: object
      required:
        - policyState
      properties:
        policyState:
          $ref: "#/components/schemas/HubEventImagePolicyState"
        url:
          type: string
          format: uri
        sourceLabel:
          type: string
        sourceUrl:
          type: string
          format: uri
        altText:
          type: string
    HubEventLinkKind:
      type: string
      enum:
        - source
        - purchase
        - ticket
        - reservation
        - content
        - video
        - map
        - custom
    HubEventLinkInput:
      type: object
      required:
        - kind
        - url
        - sortOrder
      properties:
        id:
          type: string
        kind:
          $ref: "#/components/schemas/HubEventLinkKind"
        label:
          type: string
          maxLength: 80
        url:
          type: string
          format: uri
          pattern: "^https://"
        sortOrder:
          type: integer
          minimum: 0
    HubEventLink:
      allOf:
        - $ref: "#/components/schemas/HubEventLinkInput"
        - type: object
          required:
            - id
            - createdAt
            - updatedAt
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    HubEvent:
      type: object
      required:
        - id
        - category
        - tags
        - participationMode
        - status
        - title
        - generationId
        - sourceUrl
        - sourceLabel
        - sourceType
        - notificationEligible
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            $ref: "#/components/schemas/HubEventTag"
          example: [album]
        participationMode:
          type: string
        status:
          type: string
        title:
          type: string
        summary:
          type: string
        generationId:
          type: string
        memberId:
          type: string
        sourceUrl:
          type: string
          format: uri
        sourceLabel:
          type: string
        sourceType:
          type: string
        scheduleMode:
          $ref: "#/components/schemas/HubEventScheduleMode"
        scheduleItems:
          type: array
          items:
            $ref: "#/components/schemas/HubEventScheduleItem"
        links:
          type: array
          items:
            $ref: "#/components/schemas/HubEventLink"
        announcedAt:
          type: string
          format: date-time
          nullable: true
        startsAt:
          type: string
          format: date-time
          nullable: true
        endsAt:
          type: string
          format: date-time
          nullable: true
        purchaseUrl:
          type: string
          format: uri
        ticketUrl:
          type: string
          format: uri
        venueName:
          type: string
        venueAddress:
          type: string
        image:
          $ref: "#/components/schemas/HubEventImage"
        notificationEligible:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    HubEventScheduleMode:
      type: string
      enum:
        - single_window
        - timeline
    HubEventScheduleKind:
      type: string
      enum:
        - main_window
        - announcement
        - sales_open
        - ticket_open
        - content_reveal
        - release
        - deadline
        - custom
    HubEventTimePrecision:
      type: string
      enum:
        - date
        - datetime
    HubEventScheduleItem:
      type: object
      required:
        - id
        - kind
        - label
        - startsAt
        - timePrecision
        - timezone
        - notificationEligible
        - isPrimary
        - sortOrder
      properties:
        id:
          type: string
        kind:
          $ref: "#/components/schemas/HubEventScheduleKind"
        title:
          type: string
          maxLength: 160
        label:
          type: string
          maxLength: 80
        description:
          type: string
          nullable: true
          maxLength: 2000
        startsAt:
          oneOf:
            - type: string
              format: date
            - type: string
              format: date-time
        endsAt:
          oneOf:
            - type: string
              format: date
            - type: string
              format: date-time
        timePrecision:
          $ref: "#/components/schemas/HubEventTimePrecision"
        timezone:
          type: string
          default: Asia/Seoul
        actionUrl:
          type: string
          format: uri
        sourceUrl:
          type: string
          format: uri
        sourceLabel:
          type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/HubEventLink"
        notificationEligible:
          type: boolean
        isPrimary:
          type: boolean
        sortOrder:
          type: integer
        cancelledAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    HubEventPublicationState:
      type: string
      enum:
        - draft
        - published
        - inactive
        - deleted
    HubEventTag:
      type: string
      enum: [album]
    AdminHubEventCreateRequest:
      type: object
      required:
        - title
        - category
        - participationMode
        - status
        - generationId
        - sourceUrl
        - sourceLabel
        - sourceType
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        category:
          type: string
        tags:
          type: array
          maxItems: 1
          items:
            $ref: "#/components/schemas/HubEventTag"
          example: [album]
        participationMode:
          type: string
        status:
          type: string
        generationId:
          type: string
        memberId:
          type: string
        sourceUrl:
          type: string
          format: uri
        sourceLabel:
          type: string
        sourceType:
          type: string
        scheduleMode:
          $ref: "#/components/schemas/HubEventScheduleMode"
        scheduleItems:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/HubEventScheduleItem"
        links:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/HubEventLinkInput"
        announcedAt:
          type: string
          format: date-time
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        purchaseUrl:
          type: string
          format: uri
        ticketUrl:
          type: string
          format: uri
        venueName:
          type: string
        venueAddress:
          type: string
        notificationEligible:
          type: boolean
    AdminHubEventUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/AdminHubEventCreateRequest"
    AdminHubEventRevisionRequest:
      type: object
      required:
        - expectedRevision
      properties:
        expectedRevision:
          type: integer
          minimum: 1
        changeReason:
          type: string
    AdminHubEventScheduleItemMutationRequest:
      allOf:
        - type: object
          properties:
            kind:
              $ref: "#/components/schemas/HubEventScheduleKind"
            title:
              type: string
              maxLength: 160
            label:
              type: string
            description:
              type: string
            startsAt:
              type: string
            endsAt:
              type: string
            timePrecision:
              $ref: "#/components/schemas/HubEventTimePrecision"
            timezone:
              type: string
              default: Asia/Seoul
            actionUrl:
              type: string
              format: uri
            sourceUrl:
              type: string
              format: uri
            sourceLabel:
              type: string
            links:
              type: array
              maxItems: 10
              items:
                $ref: "#/components/schemas/HubEventLinkInput"
            notificationEligible:
              type: boolean
            isPrimary:
              type: boolean
            sortOrder:
              type: integer
        - $ref: "#/components/schemas/AdminHubEventRevisionRequest"
    AdminHubEventScheduleItemCreateRequest:
      allOf:
        - $ref: "#/components/schemas/AdminHubEventScheduleItemMutationRequest"
        - type: object
          required:
            - kind
            - label
            - startsAt
            - timePrecision
            - timezone
    AdminHubEventScheduleOrderRequest:
      allOf:
        - $ref: "#/components/schemas/AdminHubEventRevisionRequest"
        - type: object
          required:
            - scheduleItemIds
          properties:
            scheduleItemIds:
              type: array
              items:
                type: string
    AdminHubEventScheduleDeleteResponse:
      type: object
      required:
        - event
        - scheduleItemId
        - deletion
      properties:
        event:
          $ref: "#/components/schemas/AdminHubEventResponse"
        scheduleItemId:
          type: string
        deletion:
          type: string
          enum: [cancelled, hard_deleted]
    AdminHubEventResponse:
      allOf:
        - $ref: "#/components/schemas/HubEvent"
        - type: object
          required:
            - publicationState
            - revision
          properties:
            publicationState:
              $ref: "#/components/schemas/HubEventPublicationState"
            publishedAt:
              type: string
              format: date-time
            cancelledAt:
              type: string
              format: date-time
            deactivatedAt:
              type: string
              format: date-time
            deletedAt:
              type: string
              format: date-time
            revision:
              type: integer
              minimum: 1
            createdBy:
              type: string
            updatedBy:
              type: string
    AdminHubEventListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminHubEventResponse"
        nextCursor:
          type: string
    AdminHubEventValidationError:
      type: object
      required:
        - field
        - reason
        - message
      properties:
        field:
          type: string
        reason:
          type: string
        message:
          type: string
    HubEventAuditLogEntry:
      type: object
      required:
        - id
        - hubEventId
        - action
        - createdAt
      properties:
        id:
          type: string
        hubEventId:
          type: string
        action:
          type: string
        actorId:
          type: string
        reason:
          type: string
        before:
          type: object
        after:
          type: object
        createdAt:
          type: string
          format: date-time
...